requirements.txt
Flask
requests
psycopg2-binary
redis
rq
python-dotenv
transformers
torch